basic_backend/requirements.txt
asgiref==3.8.1
Django==3.2.3
django-cors-headers==3.11.0
django-createsuperuserwithpassword==2.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dotenv==0.10.3
sqlparse==0.5.0
typing-extensions==4.12.0
tzdata==2024.1
